requirements.txt
colormath==3.0.0
discord.py==1.7.3
requests==2.27.1
